refactor: separate logging from main
