+++ src/main.ts
@@ -4,9 +4,8 @@ import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
 import { ValidationPipe } from '@nestjs/common';
 import * as dotenv from 'dotenv';
 import { t } from './shared/loc';
-import { LoggingService } from './core/services/logging/logging.service';
-import { AllExceptionsFilter } from './core/filters/allExceptions.filter';
 import setupGlobalAuthGuard from './auth/utils/setupGlobalAuthGuard';
+import setupGlobalLogging from './core/utils/setupGlobalLogging';
 
 dotenv.config();
 
@@ -17,19 +16,8 @@ const VERSION = process.env.VERSION || '1.0.0';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
-  const loggingService = app.get(LoggingService);
-
+  setupGlobalLogging(app);
   setupGlobalAuthGuard(app);
-  app.useGlobalFilters(new AllExceptionsFilter(loggingService));
-
-  process.on('uncaughtException', (err) => {
-    loggingService.error('Uncaught Exception', err.stack, 'Bootstrap');
-  });
-
-  process.on('unhandledRejection', (reason) => {
-    loggingService.error('Unhandled Rejection', reason as string, 'Bootstrap');
-  });
-
   app.useGlobalPipes(new ValidationPipe());
 
   const config = new DocumentBuilder()
